Clear up a loss-of-precision warning
rb_funcall2 expects an int, std::vector::size returns a larger type
depending on architecture. Match what Ruby's API expects.

+++ rice/Object.ipp
@@ -12,7 +12,7 @@ inline Rice::Object Rice::Object::
 call(Identifier id, ArgT... args) const
 {
   auto asList = this->convert_args<ArgT...>(args...);
-  return protect(rb_funcall2, value(), id, asList.size(), &asList[0]);
+  return protect(rb_funcall2, value(), id, (int)asList.size(), &asList[0]);
 }
 
 template<typename ...ArgT>
